first com

+++ src/screens/Payment.js
@@ -220,7 +220,7 @@ function Payment(props) {
 
       var val = Math.floor(1000 + Math.random() * 9000);
 
-      var taxvalue= (sum/100)*5;
+      var taxvalue= float((sum/100)*5);
  
 
       const [Field,setField] =useState(false);
